f16/scalar
